Fixed error in averageScore function

+++ Code/Server/restserver/src/main/java/data/model/Player.java
@@ -104,8 +104,8 @@ public class Player extends User implements Serializable {
 	}
 	
 	public double averageScore() {
-		int averageScore = 0;
-		int counter = 0;
+		double averageScore = 0;
+		double counter = 0;
 		for(Participation p : playedRounds) {
 			averageScore += p.getScore();
 			counter++;
